moving providers to dedicated directory
