+++ app/javascript/components/KeywordSearch.js
@@ -4,8 +4,8 @@ import InputGroup from 'react-bootstrap/lib/InputGroup'
 import Form from 'react-bootstrap/lib/Form'
 import { faSearch, faTimes } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { SearchSelectionContext } from './search/SearchSelectionProvider'
-import { StudySearchContext } from './search/StudySearchProvider'
+import { SearchSelectionContext } from 'providers/SearchSelectionProvider'
+import { StudySearchContext } from 'providers/StudySearchProvider'
 
 /**
  * Component to search using a keyword value
