+++ app/javascript/components/SearchQueryDisplay.js
@@ -1,8 +1,8 @@
 import React, { useContext } from 'react'
 import { faSearch } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { getDisplayNameForFacet } from 'components/search/SearchFacetProvider'
-import { SearchSelectionContext } from './search/SearchSelectionProvider'
+import { getDisplayNameForFacet } from 'providers/SearchFacetProvider'
+import { SearchSelectionContext } from 'providers/SearchSelectionProvider'
 import Button from 'react-bootstrap/lib/Button'
