+++ app/javascript/providers/SearchFacetProvider.js
