+++ app/models/user.rb
@@ -98,9 +98,9 @@ class User
 
   DEFAULT_FEATURE_FLAGS = {
     # whether the home page uses React and the new search API
-    "advanced_search" => false,
+    "advanced_search" => true,
     # whether the facet search controls are shown
-    "faceted_search" => false,
+    "faceted_search" => true,
     # show covid-19 tab on homepage
     "covid19_page" => false
   }
