+++ test/js/search/facet-control.test.js
@@ -6,7 +6,7 @@ import * as Reach from '@reach/router'
 const fetch = require('node-fetch');
 
 import FacetControl from 'components/FacetControl';
-import { PropsStudySearchProvider } from 'components/search/StudySearchProvider';
+import { PropsStudySearchProvider } from 'providers/StudySearchProvider';
 import * as ScpAPI from 'lib/scp-api'
 
 const speciesFacet = {
