+++ test/js/search/more-facets.test.js
@@ -6,7 +6,7 @@ import * as Reach from '@reach/router'
 const fetch = require('node-fetch');
 
 import MoreFacetsButton from 'components/MoreFacetsButton';
-import { PropsStudySearchProvider } from 'components/search/StudySearchProvider';
+import { PropsStudySearchProvider } from 'providers/StudySearchProvider';
 
 import * as ScpAPI from 'lib/scp-api'
